Broke parser into smaller classes, rubocop seeds
