+++ app/lib/parser.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+require_relative './exceptions/invalid_environment_variable'
+require_relative './exceptions/invalid_payload_key'
+require_relative './parser/**'
+
+module Parser
+end
