+++ app/lib/request.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+require_relative './request/**'
+
+module Request
+end
