+++ app/lib/request/deconstructor.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+module Request
+  # Deconstructs the request & response from us & service B. Replaces
+  # any headers that contain environment variables with placeholder values
+  # to prevent data leakage.
+  class Deconstructor
+  end
+end
