+++ app/lib/request/handler.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+module Request
+  class Handler
+  end
+end
